requirements.txt
bleach==1.5.0
cycler==0.10.0
Django==1.11.6
h5py==2.7.1
html5lib==0.9999999
image==1.5.16
Keras==2.0.8
keras-tqdm==2.0.1
Markdown==2.6.9
matplotlib==2.0.2
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
sklearn==0.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
tqdm==4.19.1.post1
Werkzeug==0.12.2
